SQL/api/requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
aniso8601==9.0.1
asn1crypto==1.4.0
async-timeout==4.0.1
asynctest==0.13.0
attrs==21.2.0
cachetools==4.2.4
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
click==8.0.3
cloud-sql-python-connector==0.4.2
cryptography==36.0.0
Flask==2.0.2
Flask-Cors==3.0.10
flask-restx==0.5.1
frozenlist==1.2.0
google-api-core==2.2.2
google-api-python-client==2.31.0
google-auth==2.3.3
google-auth-httplib2==0.1.0
googleapis-common-protos==1.53.0
greenlet==1.1.2
httplib2==0.20.2
idna==3.3
importlib-metadata==4.8.2
importlib-resources==5.4.0
itsdangerous==2.0.1
Jinja2==3.0.3
jsonschema==4.2.1
MarkupSafe==2.0.1
multidict==5.2.0
pg8000==1.22.0
protobuf==3.19.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyMySQL==1.0.2
pyOpenSSL==21.0.0
pyparsing==3.0.6
pyrsistent==0.18.0
python-tds==1.11.0
pytz==2021.3
requests==2.26.0
rsa==4.8
scramp==1.4.1
six==1.16.0
SQLAlchemy==1.4.27
sqlalchemy-pytds==0.3.3
typing-extensions==4.0.0
uritemplate==4.1.1
urllib3==1.26.7
Werkzeug==2.0.2
yarl==1.7.2
zipp==3.6.0
